+++ lib/eve_online.rb
@@ -54,6 +54,7 @@ require 'eve_online/esi/races'
 require 'eve_online/esi/bloodlines'
 require 'eve_online/esi/factions'
 require 'eve_online/esi/ancestries'
+require 'eve_online/esi/universe_types'
 
 require 'eve_online/esi/models/base'
 require 'eve_online/esi/models/loyalty_point'