+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 **master**
 
 * Add `EveOnline::ESI::UniverseTypes`
+* Add `EveOnline::ESI::UniverseType`
+* Add `EveOnline::ESI::Models::Type`
 
 **v0.15.0**